Updates addListener to return a closure of the unsubscribe function
If an app wanted to register several listeners to the same string
token event, removing the correct event from the list seems tricky.

I think you could filter the list for the callback (and make them
pass you the same callback they did on subscription) and then take
it out of the array, but this seemed more straightforward. The user
has the unsubscribe to use whenever they want, they just have to
capture it.

Trying this implementation for now. Passes tests.