Adds a removeAllListeners method and passing test
I looked into the most efficient way to clear an array -- I was
worried that overwriting the array with a new one might cause problems
because references to the old array might be hanging about. Setting
the length to 0 is destructive of the current eventName array, and
that seemed like a safer approach.

Successfully adds events to array then removesAll.