Completes removeListener function, passes test
The EE API now has two methods for unsubscribing: you can capture the returned
unsubscribe function when you register as a listener and invoke that later, OR you
can call removeListener with a cb function and it find the first instance of that
handler in the and eventName array remove it. This struck me as a kinda weird API,
but it's what the Node EventEmitter
does, so I'm using it as a model.