Adds once method and passing test for EE class
The once takes a handler and wraps it in a closure which first
removes the handler from the array, and then calls it with how every
many args were passed to the parent wrapper.

Required debugging to get this to work because I was trying to
remove the handler function instead of the wrapper.

Uses the built-in remove method, which isn't great because it removes
the first match, rather than the exact handler. Not sure how to fix this.