Refactors tests to be agnostic of implementation details
The first round of TDD testing was for me to prove that the EE
worked as expected -- it used internals of the data structures for
testing (checking array lengths, membership in hash, etc). Ideally
tests should be agnostic about implementation, so this refactors
all tests to simply use the EE methods with locally closed over
counters to demonstrate it works. All tests passing.